# openapi.yaml
openapi: 3.1.0
info:
  title: League Table API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  description: |
    The backend the League Table frontend (`frontend/services/httpApi.js`) talks to.

    Scoring, the commitments adjustment, ranking and tie-breaking all run on the server;
    the browser only renders the result. A student's raw score is weighted by criterion,
    multiplied by a capped factor from the weekly hours they report and approved by the
    instructor, and clipped at 100. Classmates see only the final adjusted score and rank.

    **Conventions**
    - Every field, query parameter and path parameter is `snake_case`, matching the
      data model in `_docs/spec.md` and FastAPI's defaults. The frontend's services
      layer translates its own JS option names (`weekId`, `windowMode`, ...) at the
      HTTP boundary.
    - Timestamps (`last_updated`, `submitted_at`, ...) are ISO 8601 UTC strings, or
      `null` before any successful read.
    - Errors carry a human-readable `detail` message and an HTTP status; the client
      shows the message to the user.

    **Authentication**
    Every endpoint except `POST /auth/login` and `GET /demo/accounts` needs a session
    cookie. Roles are enforced by the server, never the browser. An `instructor` can do
    everything; a `student` may read the full ranking but only their own breakdown and
    commitments, and may not change settings or approve anything. Sign-in is by access
...
servers:
...
security:
  - sessionCookie: []
paths:
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Sign in with an access code or the instructor passcode
      description: |
        A student's personal access code or the instructor's passcode, exchanged for a
        session. The service stores only a hash of each student's code. With demo data,
        the codes from `GET /demo/accounts` work here. A wrong code returns 401.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Signed in. The session cookie is set on the response.
          headers:
            Set-Cookie:
              description: HttpOnly session cookie.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentAccount
      summary: The signed-in account
      description: Lets the frontend learn the caller's role and, for students, their `student_id`.
      responses:
        '200':
          description: The current account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: End the session
      responses:
        '204':
          description: Session ended and cookie cleared.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /demo/accounts:
    get:
      tags: [Auth]
      operationId: listDemoAccounts
      summary: Demo accounts and their codes (demo data only)
      description: |
        Lists the seeded demo instructor and students so sign-in and role rules can be
        tried without real people, and includes their access codes for
        `POST /auth/login`. Exists only while the toy adapter is the active data
        source; any other source returns 404. It is never part of the class
        bootstrap.
      security: []
      responses:
        '200':
          description: Demo accounts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DemoAccount'
        '404':
          description: The active data source is not demo data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /classes/{class_id}:
    get:
      tags: [Class]
      operationId: getClass
      summary: Everything the page needs to start
      description: |
        The frontend's `getBootstrap`. Returns the class, its weeks and criteria, the
        saved weights (or each criterion's `default_weight` if none are saved), and
        the data-source status, in one call.
      parameters:
        - $ref: '#/components/parameters/ClassId'
      responses:
        '200':
          description: Class bootstrap.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassBootstrap'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/SourceUnavailable'
  /classes/{class_id}/weeks:
    get:
      tags: [Class]
      operationId: listWeeks
      summary: Weeks of the term, in order
      parameters:
        - $ref: '#/components/parameters/ClassId'
      responses:
        '200':
          description: Weeks, oldest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Week'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/SourceUnavailable'
  /classes/{class_id}/criteria:
    get:
      tags: [Class]
      operationId: listCriteria
      summary: Ranking criteria, in display order
      parameters:
        - $ref: '#/components/parameters/ClassId'
      responses:
        '200':
          description: Criteria ordered by `sort_order`.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Criterion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/SourceUnavailable'
  /classes/{class_id}/settings:
    get:
      tags: [Class]
      operationId: getSettings
      summary: Criterion weights and adjustment parameters (instructor only)
      description: Weights fall back to each criterion's `default_weight` when nothing has been saved.
      parameters:
        - $ref: '#/components/parameters/ClassId'
      responses:
        '200':
          description: The current settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/SourceUnavailable'
    put:
      tags: [Class]
      operationId: saveSettings
      summary: Save criterion weights and adjustment parameters (instructor only)
      description: |
        Every field is optional; what is left out keeps its value. Criterion weights are
        rescaled to total 100, preserving proportions, and criteria left out are stored as
        zero. Adjustment fields are merged into the current ones and range-checked. Rankings
        are computed on the fly, so a change re-ranks every week. Adjustment changes are
        written to the change log.
      parameters:
        - $ref: '#/components/parameters/ClassId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveSettingsRequest'
      responses:
        '200':
          description: The stored settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/SourceUnavailable'
  /classes/{class_id}/explainer:
    get:
      tags: [Class]
      operationId: getExplainer
      summary: The scoring formula and its current parameters
      description: |
        The same for everyone: the formula, the current type weights, rate and cap, a
        factor table and worked examples. It holds no one's hours.
      parameters:
        - $ref: '#/components/parameters/ClassId'
      responses:
        '200':
          description: The explainer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Explainer'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: HttpOnly session cookie set by `/auth/login`.
  parameters:
    ClassId:
      name: class_id
      in: path
      required: true
      schema:
        type: string
      example: c1
...
  responses:
    Unauthorized:
      description: No valid session.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The caller's role does not allow this.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: A student may only open their own breakdown.
    NotFound:
      description: Unknown class.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: |
        A parameter or body field is missing or malformed. Uses FastAPI's default 422
        body, where `detail` is a list of per-field problems.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    SourceUnavailable:
      description: |
        The data source did not respond and there is no earlier snapshot to serve.
        When a snapshot exists the server serves it with `stale: true` instead.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: The data source did not respond.
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Message safe to show to the user.
    ValidationError:
      type: object
      required: [detail]
      properties:
        detail:
          type: array
          items:
            type: object
            required: [loc, msg, type]
            properties:
              loc:
                type: array
                description: Where the problem is, e.g. `["query", "week"]`.
                items:
                  type: [string, integer]
              msg:
                type: string
              type:
                type: string
    Role:
      type: string
      enum: [instructor, student]
    Account:
      type: object
      required: [id, role, student_id, external_id]
      properties:
        id:
          type: string
          example: a2
        role:
          $ref: '#/components/schemas/Role'
        student_id:
          type: [string, 'null']
          description: The student this account belongs to; `null` for the instructor.
          example: s01
        external_id:
          type: string
          description: ID from an outside identity system, so sign-in can be matched later.
          example: demo:amara
    DemoAccount:
      description: |
        A seeded demo account with its access code. These codes are public by design and
        must never be accepted by a deployment using real student data.
      allOf:
        - $ref: '#/components/schemas/Account'
        - type: object
          required: [name, code]
          properties:
            name:
              type: string
              example: Amara Okonkwo
            code:
              type: string
              example: demo-amara
    LoginRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          format: password
          minLength: 1
          description: A student's access code, or the instructor's passcode.
    Class:
      type: object
      required: [id, external_id, name, term_id]
      properties:
        id:
          type: string
          example: c1
        external_id:
          type: string
          example: demo:phys101-b
        name:
          type: string
          example: Introductory Physics · Group B
        term_id:
          type: string
          example: t1
    Week:
      type: object
      required: [id, term_id, week_number, start_date, end_date]
      properties:
        id:
          type: string
          example: w1
        term_id:
          type: string
          example: t1
        week_number:
          type: integer
          example: 1
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
    Criterion:
      type: object
      required: [id, class_id, key, label, unit, default_weight, sort_order]
      properties:
        id:
          type: string
        class_id:
          type: string
        key:
          type: string
          description: Stable identifier used in `criteria`, `weights` and score parts.
          example: homework
        label:
          type: string
          example: Homework
        unit:
          type: string
          example: tasks on time
        default_weight:
          type: number
          minimum: 0
          example: 25
        sort_order:
          type: integer
    Weights:
      type: object
      description: Criterion key to weight. Values are non-negative.
      additionalProperties:
        type: number
        minimum: 0
      example: { homework: 25, attendance: 20, participation: 20, project: 25, quizzes: 10 }
    AdjustmentSettings:
      type: object
      description: |
        The parameters of the adjustment. Weighted hours are `H = sum(a_k × h_k)`, the factor is
        `min(1 + rate × H, cap)`, and the adjusted score is `min(100, raw × factor)`.
      required: [type_weights, rate, cap, flag_hours]
      properties:
        type_weights:
          $ref: '#/components/schemas/Weights'
        rate:
          type: number
          minimum: 0
          maximum: 1
          description: Factor added per weighted hour.
          example: 0.01
        cap:
          type: number
          minimum: 1
          maximum: 3
          description: The most the factor can reach.
          example: 1.25
        flag_hours:
          type: number
          minimum: 0
          description: A weekly entry differing from the baseline by more total hours than this is flagged.
          example: 10
    Settings:
      type: object
      required: [weights, adjustment]
      properties:
        weights:
          $ref: '#/components/schemas/Weights'
        adjustment:
          $ref: '#/components/schemas/AdjustmentSettings'
    SaveSettingsRequest:
      type: object
      properties:
        weights:
          $ref: '#/components/schemas/Weights'
        adjustment:
          type: object
          description: Only the fields given are changed.
          properties:
            type_weights:
              $ref: '#/components/schemas/Weights'
            rate:
              type: number
            cap:
              type: number
            flag_hours:
              type: number
    SourceInfo:
      type: object
      required: [kind, label, demo]
      properties:
        kind:
          type: string
          description: Which data-source adapter is active (`toy` for seeded demo data).
          example: toy
        label:
          type: string
          example: Demo data
        demo:
          type: boolean
          description: True when the data is seeded demo data, so the UI can say so.
    Issue:
      type: object
      description: A problem found while validating the source data. The table is still served.
      required: [level, where, message]
      properties:
        level:
          type: string
          enum: [error, warning]
        where:
          type: string
          description: ID of the offending entry.
          example: e12
        message:
          type: string
          example: Unknown student ID "s99"
    ClassBootstrap:
      type: object
      required: [class, weeks, criteria, weights, tie_breakers, source, last_updated, issues]
      properties:
        class:
          $ref: '#/components/schemas/Class'
        weeks:
          type: array
          items:
            $ref: '#/components/schemas/Week'
        criteria:
          type: array
          items:
            $ref: '#/components/schemas/Criterion'
        weights:
          $ref: '#/components/schemas/Weights'
        tie_breakers:
          type: array
          description: Labels of the criteria that order students still level after the raw score, in priority order.
          items:
            type: string
          example: [Attendance, Homework]
        source:
          $ref: '#/components/schemas/SourceInfo'
        last_updated:
          type: string
          format: date-time
          description: When the source was last read successfully (ISO 8601, UTC).
        issues:
          type: array
          items:
            $ref: '#/components/schemas/Issue'
...
    Hours:
      type: object
      description: Hours per week by commitment type key, such as `work`, `childcare` and `eldercare`.
      additionalProperties:
        type: number
      example: { work: 12, childcare: 6, eldercare: 0 }
    CommitmentType:
      type: object
      required: [key, label]
      properties:
        key:
          type: string
          example: childcare
        label:
          type: string
          example: Child care
    HourLimits:
      type: object
      required: [step, per_type_max, total_max]
      properties:
        step:
          type: number
          example: 0.5
        per_type_max:
          type: number
          example: 80
        total_max:
          type: number
          example: 120
...
    Explainer:
      type: object
      description: The formula and current parameters, with worked examples. It holds no personal data.
      required: [formula, types, type_weights, rate, cap, limits, factor_table, examples, rules]
      properties:
        formula:
          type: array
          items:
            type: string
        types:
          type: array
          items:
            $ref: '#/components/schemas/CommitmentType'
        type_weights:
          $ref: '#/components/schemas/Weights'
        rate:
          type: number
        cap:
          type: number
        limits:
          $ref: '#/components/schemas/HourLimits'
        factor_table:
          type: array
          items:
            type: object
            required: [weighted_hours, factor]
            properties:
              weighted_hours:
                type: number
              factor:
                type: number
        examples:
          type: array
          items:
            type: object
            required: [title, raw_score, hours, weighted_hours, factor, adjusted_score, capped, steps]
            properties:
              title:
                type: string
              raw_score:
                type: number
              hours:
                $ref: '#/components/schemas/Hours'
              weighted_hours:
                type: number
              factor:
                type: number
              adjusted_score:
                type: number
              capped:
                type: boolean
              steps:
                type: array
                items:
                  type: string
        rules:
          type: array
          items:
            type: string